+++ bindings/screamer_bindings.cpp
@@ -2,11 +2,23 @@
 #include <pybind11/numpy.h>
 #include "screamer/lag.h"
 #include "screamer/diff.h"
+#include "screamer/fracdiff.h"
 #include "screamer/rolling_sum.h"
+#include "screamer/rolling_mean.h"
+#include "screamer/rolling_std.h"
+#include "screamer/rolling_var.h"
+#include "screamer/rolling_skew.h"
+#include "screamer/rolling_kurt.h"
+#include "screamer/rolling_zscore.h"
+#include "screamer/rolling_min.h"
+#include "screamer/rolling_max.h"
+#include "screamer/rolling_median.h"
+#include "screamer/rolling_median2.h"
 #include "screamer/return.h"
 #include "screamer/log_return.h"
-#include "screamer/sma.h"
-#include "screamer/fracdiff.h"
+#include "screamer/ewma.h"
+#include "screamer/ew_var.h"
+#include "screamer/ew_std.h"
 
 namespace py = pybind11;
 
@@ -23,7 +35,8 @@ PYBIND11_MODULE(screamer_bindings, m) {
              "This applies a lag of `d` to the stream of input values `X`.")
         .def("reset", &screamer::Lag::reset, "Reset the Lag indicator to its initial state.")
         .def("transform", &screamer::Lag::transform, py::arg("input_array"),
-             "Apply the lag transformation to a NumPy array.");
+             "Apply the lag transformation to a NumPy array.")
+        .doc() = "Lag indicator, a delayed version of values in a sequence.";
 
     py::class_<screamer::Diff>(m, "Diff")
         .def(py::init<int>(), py::arg("delay"),
@@ -36,7 +49,8 @@ PYBIND11_MODULE(screamer_bindings, m) {
              "This applies a diff of `d` to the stream of input values `X`.")
         .def("reset", &screamer::Diff::reset, "Reset the diff indicator to its initial state.")
         .def("transform", &screamer::Diff::transform, py::arg("input_array"),
-             "Apply the diff transformation to a NumPy array.");
+             "Apply the diff transformation to a NumPy array.")
+        .doc() = "Diff indicator, relative compared to past values.";
 
     py::class_<screamer::RollingSum>(m, "RollingSum")
         .def(py::init<int>(), py::arg("window_size"),
@@ -49,7 +63,78 @@ PYBIND11_MODULE(screamer_bindings, m) {
              "This applies a rolling sum the stream of input values `X`.")
         .def("reset", &screamer::RollingSum::reset, "Reset the rolling sum indicator to its initial state.")
         .def("transform", &screamer::RollingSum::transform, py::arg("input_array"),
-             "Apply the rolling sum transformation to a NumPy array.");
+             "Apply the rolling sum transformation to a NumPy array.")
+        .doc() = "Rolling sum, the sum of the most recent values in a sequence.";
+
+    py::class_<screamer::RollingMean>(m, "RollingMean")
+        .def(py::init<int>(), py::arg("window_size"),
+             "Initialize the simple moving average indicator.\n\n"
+             ":param window_size: The window_size parameter.")
+        .def("__call__", &screamer::RollingMean::operator(), py::arg("value"),
+             "Update and return the simple moving average.\n\n"
+             ".. math::\n"
+             "    f(X[i]) = \\sum_{i=0}^{w-1}X[i-w] / w\n\n"
+             "This applies a simple moving average to the stream of input values `X`.")
+        .def("reset", &screamer::RollingMean::reset, "Reset the simple moving average indicator to its initial state.")
+        .def("transform", &screamer::RollingMean::transform, py::arg("input_array"),
+             "Apply the simple moving average transformation to a NumPy array.")
+        .doc() = "Simple moving average, the average of the most recent values in a sequence.";
+
+
+    py::class_<screamer::RollingStd>(m, "RollingStd")
+        .def(py::init<int>(), py::arg("window_size"),
+             "Initialize the rolling standard deviation indicator.\n\n"
+             ":param window_size: The window_size parameter.")
+        .def("__call__", &screamer::RollingStd::operator(), py::arg("value"),
+             "This applies a rolling standard deviation the stream of input values `X`.")
+        .def("reset", &screamer::RollingStd::reset, "Reset the indicator to its initial state.")
+        .def("transform", &screamer::RollingStd::transform, py::arg("input_array"),
+             "Apply the indicator to a NumPy array.")
+        .doc() = "Rolling standard deviation, the standard deviation of the most recent values in a sequence.";
+
+    py::class_<screamer::RollingVar>(m, "RollingVar")
+        .def(py::init<int>(), py::arg("window_size"),
+             "Initialize the rolling variance indicator.\n\n"
+             ":param window_size: The window_size parameter.")
+        .def("__call__", &screamer::RollingVar::operator(), py::arg("value"),
+             "This applies a rolling variance to the stream of input values `X`.")
+        .def("reset", &screamer::RollingVar::reset, "Reset the indicator to its initial state.")
+        .def("transform", &screamer::RollingVar::transform, py::arg("input_array"),
+             "Apply the indicator to a NumPy array.")
+        .doc() = "Rolling variance, the variance of the most recent values in a sequence.";
+
+    py::class_<screamer::RollingSkew>(m, "RollingSkew")
+        .def(py::init<int>(), py::arg("window_size"),
+             "Initialize the rolling skew indicator.\n\n"
+             ":param window_size: The window_size parameter.")
+        .def("__call__", &screamer::RollingSkew::operator(), py::arg("value"),
+             "This applies a rolling skew to the stream of input values `X`.")
+        .def("reset", &screamer::RollingSkew::reset, "Reset the indicator to its initial state.")
+        .def("transform", &screamer::RollingSkew::transform, py::arg("input_array"),
+             "Apply the indicator to a NumPy array.")
+        .doc() = "Rolling skew, the skew of the most recent values in a sequence.";
+
+    py::class_<screamer::RollingKurt>(m, "RollingKurt")
+        .def(py::init<int>(), py::arg("window_size"),
+             "Initialize the rolling kurtosis indicator.\n\n"
+             ":param window_size: The window_size parameter.")
+        .def("__call__", &screamer::RollingKurt::operator(), py::arg("value"),
+             "This applies a rolling kurtosis to the stream of input values `X`.")
+        .def("reset", &screamer::RollingKurt::reset, "Reset the indicator to its initial state.")
+        .def("transform", &screamer::RollingKurt::transform, py::arg("input_array"),
+             "Apply the indicator to a NumPy array.")
+        .doc() = "Rolling kurtosis, the kurtosis of the most recent values in a sequence.";
+
+    py::class_<screamer::RollingZscore>(m, "RollingZscore")
+        .def(py::init<int>(), py::arg("window_size"),
+             "Initialize the rolling zscore indicator.\n\n"
+             ":param window_size: The window_size parameter.")
+        .def("__call__", &screamer::RollingZscore::operator(), py::arg("value"),
+             "This applies a rolling zscore the stream of input values `X`.")
+        .def("reset", &screamer::RollingZscore::reset, "Reset the indicator to its initial state.")
+        .def("transform", &screamer::RollingZscore::transform, py::arg("input_array"),
+             "Apply the indicator to a NumPy array.")
+        .doc() = "Rolling zscore, the zscore of the most recent values in a sequence.";
 
     py::class_<screamer::Return>(m, "Return")
         .def(py::init<int>(), py::arg("delay"),
@@ -62,7 +147,8 @@ PYBIND11_MODULE(screamer_bindings, m) {
              "This applies a simple return to the stream of input values `X`.")
         .def("reset", &screamer::Return::reset, "Reset the simple return indicator to its initial state.")
         .def("transform", &screamer::Return::transform, py::arg("input_array"),
-             "Apply the simple return transformation to a NumPy array.");
+             "Apply the simple return transformation to a NumPy array.")
+        .doc() = "Simple return, the percentage change compared to a previous value";
 
     py::class_<screamer::LogReturn>(m, "LogReturn")
         .def(py::init<int>(), py::arg("delay"),
@@ -75,33 +161,109 @@ PYBIND11_MODULE(screamer_bindings, m) {
              "This applies a log return to the stream of input values `X`.")
         .def("reset", &screamer::LogReturn::reset, "Reset the log return indicator to its initial state.")
         .def("transform", &screamer::LogReturn::transform, py::arg("input_array"),
-             "Apply the log return transformation to a NumPy array.");
+             "Apply the log return transformation to a NumPy array.")
+        .doc() = "Log return, the logarithmic change compared to a previous value.";
 
-    py::class_<screamer::SMA>(m, "SMA")
-        .def(py::init<int>(), py::arg("window_size"),
-             "Initialize the simple moving average indicator.\n\n"
-             ":param window_size: The window_size parameter.")
-        .def("__call__", &screamer::SMA::operator(), py::arg("value"),
+    py::class_<screamer::EWMA>(m, "EWMA")
+        .def(py::init<int>(), py::arg("weight"),
+             "Initialize the exponentially weighted moving average indicator.\n\n"
+             ":param weight: Weight of new values.")
+        .def("__call__", &screamer::EWMA::operator(), py::arg("value"),
              "Update and return the simple moving average.\n\n"
              ".. math::\n"
-             "    f(X[i]) = \\sum_{i=0}^{w-1}X[i-w] / w\n\n"
-             "This applies a simple moving average to the stream of input values `X`.")
-        .def("reset", &screamer::SMA::reset, "Reset the simple moving average indicator to its initial state.")
-        .def("transform", &screamer::SMA::transform, py::arg("input_array"),
-             "Apply the simple moving average transformation to a NumPy array.");
+             "    f(X[i]) = w * X[i] + (1 - w) * X[i-1]\n\n"
+             "This applies an exponentially weighted moving average to the stream of input values `X`.")
+        .def("reset", &screamer::EWMA::reset, "Reset the exponentially weighted moving average indicator to its initial state.")
+        .def("transform", &screamer::EWMA::transform, py::arg("input_array"),
+             "Apply the exponentially weighted moving average transformation to a NumPy array.")
+        .doc() = "Exponentially weighted moving average ofvalues in a sequence.";
 
-     py::class_<screamer::FracDiff>(m, "FracDiff")
-          .def(py::init<double, int, double>(), py::arg("frac_order"), py::arg("window_size"), py::arg("threshold")=1e-5,
-                "Initialize the fractional difference indicator.\n\n"
-                ":param frac_order: The fractional order parameter, d.\n"
-                ":param window_size: The window_size parameter.\n"
-                ":param threshold: The threshold parameter.")
-          .def("__call__", &screamer::FracDiff::operator(), py::arg("value"),
-                "Update and return the fractionally differenced value.\n\n"
-                ".. math::\n"
-                "    f(X[i]) = \\sum_{i=0}^{w-1}w[i] * X[i-w]\n\n"
-                "This applies a fractional difference to the stream of input values `X`.")
-          .def("transform", &screamer::FracDiff::transform, py::arg("input_array"),
-                "Apply the fractional difference transformation to a NumPy array.");
+    py::class_<screamer::EwStd>(m, "EwStd")
+        .def(py::init<int>(), py::arg("weight"),
+             "Initialize the exponentially weighted standard deviation indicator.\n\n"
+             ":param weight: Weight of new values.")
+        .def("__call__", &screamer::EwStd::operator(), py::arg("value"),
+             "Update and return exponentially weighted standard deviation indicator.")
+        .def("reset", &screamer::EwStd::reset, 
+        "Reset the exponentially weighted moving standard deviation indicator to its initial state.")
+        .def("transform", &screamer::EwStd::transform, py::arg("input_array"),
+             "Apply the exponentially weighted standard deviation transformation to a NumPy array.")
+        .doc() = "Exponentially weighted standard deviation of values in a sequence.";
+
+    py::class_<screamer::EwVar>(m, "EwVar")
+        .def(py::init<int>(), py::arg("weight"),
+             "Initialize the exponentially weighted variance indicator.\n\n"
+             ":param weight: Weight of new values.")
+        .def("__call__", &screamer::EwVar::operator(), py::arg("value"),
+             "Update and return exponentially weighted variance indicator.")
+        .def("reset", &screamer::EwVar::reset, 
+        "Reset the exponentially weighted moving standard deviation indicator to its initial state.")
+        .def("transform", &screamer::EwVar::transform, py::arg("input_array"),
+             "Apply the exponentially weighted variance transformation to a NumPy array.")
+        .doc() = "Exponentially weighted variance of values in a sequence.";
+
+    py::class_<screamer::RollingMin>(m, "RollingMin")
+        .def(py::init<int>(), py::arg("window_size"),
+             "Initialize the rolling min indicator.\n\n"
+             ":param window_size: The window_size parameter.")
+        .def("__call__", &screamer::RollingMin::operator(), py::arg("value"),
+             "Update and return the rolling min.")
+        .def("reset", &screamer::RollingMin::reset, "Reset the indicator to its initial state.")
+        .def("transform", &screamer::RollingMin::transform, py::arg("input_array"),
+             "Apply the indicator to a NumPy array.")
+        .doc() = "Rolling min, the minimum value of the most recent values in a sequence.";
+
+    py::class_<screamer::RollingMax>(m, "RollingMax")
+        .def(py::init<int>(), py::arg("window_size"),
+             "Initialize the rolling max indicator.\n\n"
+             ":param window_size: The window_size parameter.")
+        .def("__call__", &screamer::RollingMax::operator(), py::arg("value"),
+             "Update and return the rolling max.")
+        .def("reset", &screamer::RollingMax::reset, "Reset the indicator to its initial state.")
+        .def("transform", &screamer::RollingMax::transform, py::arg("input_array"),
+             "Apply the indicator to a NumPy array.")
+        .doc() = "Rolling max, the max value of the most recent values in a sequence.";
+
+    py::class_<screamer::RollingMedian>(m, "RollingMedian")
+        .def(py::init<int>(), py::arg("window_size"),
+             "Initialize the rolling median indicator.\n\n"
+             ":param window_size: The window_size parameter.")
+        .def("__call__", &screamer::RollingMedian::operator(), py::arg("value"),
+             "Update and return the rolling median.")
+        .def("reset", &screamer::RollingMedian::reset, "Reset the indicator to its initial state.")
+        .def("transform", &screamer::RollingMedian::transform, py::arg("input_array"),
+             "Apply the indicator to a NumPy array.")
+        .doc() = "Rolling median, the median value of the most recent values in a sequence.";
+
+
+    py::class_<screamer::RollingMedian2>(m, "RollingMedian2")
+        .def(py::init<int>(), py::arg("window_size"),
+             "Initialize the rolling median indicator.\n\n"
+             ":param window_size: The window_size parameter.")
+        .def("__call__", &screamer::RollingMedian2::operator(), py::arg("value"),
+             "Update and return the rolling median.")
+        .def("reset", &screamer::RollingMedian2::reset, "Reset the indicator to its initial state.")
+        .def("transform", &screamer::RollingMedian2::transform, py::arg("input_array"),
+             "Apply the indicator to a NumPy array.")
+        .doc() = "Rolling median, the median value of the most recent values in a sequence.";
+
+     
+ 	py::class_<screamer::FracDiff>(m, "FracDiff")
+      	.def(py::init<double, int, double>(), py::arg("frac_order"), py::arg("window_size"), py::arg("threshold")=1e-5,
+            	"Initialize the fractional difference indicator.\n\n"
+            	":param frac_order: The fractional order parameter, d.\n"
+            	":param window_size: The window_size parameter.\n"
+            	":param threshold: The threshold parameter.")
+      	.def("__call__", &screamer::FracDiff::operator(), py::arg("value"),
+            	"Update and return the fractionally differenced value.\n\n"
+            	".. math::\n"
+            	"	f(X[i]) = \\sum_{j=0}^{w-1} w[j] * X[i-j]\n\n"
+            	"This applies a fractional difference to the stream of input values `X`.")
+      	.def("transform", &screamer::FracDiff::transform, py::arg("input_array"),
+            	"Apply the fractional difference transformation to a NumPy array.")
+      	.def("reset", &screamer::FracDiff::reset, "Reset the fractional difference indicator to its initial state.")
+      	.doc() = "Fractional difference, the weighted sum of the most recent values in a sequence.";
+
+
 
 }