+++ include/screamer/rolling_sum.h
@@ -1,40 +1,24 @@
 #ifndef SCREAMER_ROLLINGSUM_H
 #define SCREAMER_ROLLINGSUM_H
 
-#include <vector>
-#include <stdexcept>
 #include <limits>
 #include <pybind11/pybind11.h>
 #include <pybind11/numpy.h>
 #include <screamer/transforms.h>
+#include <screamer/buffer.h>
 
 namespace py = pybind11; // Alias for pybind11 namespace
 
 namespace screamer {
 
-// RollingSum, skip NaNs
 class RollingSum {
 public:
-    // Constructor with an integer windows size (N)
-    RollingSum(int N)
-        : index(0), N(N), sum(0.0)
-    {
-        if (N < 1) {
-            throw std::invalid_argument("Window size must be an integer >= 1.");
-        }
-        buffer.resize(N, std::numeric_limits<double>::quiet_NaN());    
-    }
 
-    // Operator to be used as the function call operator for applying the lag
-    double operator()(double newValue) 
+    RollingSum(int N) : sum(0), buffer(N, std::numeric_limits<double>::quiet_NaN()) {}
+    
+    double operator()(const double newValue) 
     {
-        double oldValue = buffer[index];
-        buffer[index] = newValue;
-
-        index++;
-        if (index == N) {
-            index = 0;
-        }
+        double oldValue = buffer.append(newValue);
 
         if (!std::isnan(oldValue)) {
             sum -= oldValue;
@@ -48,24 +32,19 @@ public:
 
     }
 
-    // reset the internal state
     void reset() 
     {
-        std::fill(buffer.begin(), buffer.end(), std::numeric_limits<double>::quiet_NaN());
-        index = 0;
+        buffer.reset(std::numeric_limits<double>::quiet_NaN());
         sum = 0.0;
     }
 
-    // Method to transform a NumPy array, applying the lag transformation
-    py::array_t<double> transform(py::array_t<double> input_array) 
+    py::array_t<double> transform(const py::array_t<const double> input_array) 
     {
         return transform_1(*this, input_array);
     }
 
 private:
-    size_t index; // Tracks the current position in the buffer
-    size_t N;     // The window size value
-    std::vector<double> buffer; // Used as circular buffer for storing lagged values
+    FixedSizeBuffer buffer;
     double sum;
 };